Moved AuthUser and AuthRole to a sub folder for security entities
