+++ MessagingApp.Infrastructure/Data/Entities/Security/AuthRole.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Identity;
 
-namespace MessagingApp.Infrastructure.Data.Entities;
+namespace MessagingApp.Infrastructure.Data.Entities.Security;
 
 public class AuthRole : IdentityRole<Guid>
 {
