+++ MessagingApp.Infrastructure/Data/Entities/Security/AuthUser.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Identity;
 
-namespace MessagingApp.Infrastructure.Data.Entities;
+namespace MessagingApp.Infrastructure.Data.Entities.Security;
 
 public class AuthUser : IdentityUser<Guid>
 {
