+++ MessagingApp.Infrastructure/Data/Repositories/UserRepository.cs
@@ -4,6 +4,7 @@ using MessagingApp.Application.Common.Interfaces.Repositories;
 using MessagingApp.Domain.Entities;
 using MessagingApp.Infrastructure.Data.Contexts;
 using MessagingApp.Infrastructure.Data.Entities;
+using MessagingApp.Infrastructure.Data.Entities.Security;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
