+++ MessagingApp.Infrastructure/InfrastructureDependencyInjection.cs
@@ -1,7 +1,7 @@
 ﻿using MessagingApp.Application.Common.Interfaces.Repositories;
 using MessagingApp.Application.Common.Interfaces.Services;
 using MessagingApp.Infrastructure.Data.Contexts;
-using MessagingApp.Infrastructure.Data.Entities;
+using MessagingApp.Infrastructure.Data.Entities.Security;
 using MessagingApp.Infrastructure.Data.Repositories;
 using MessagingApp.Infrastructure.Services;
 using Microsoft.AspNetCore.Identity;
